requirement.txt
fastapi
uvicorn
opensearch-py
pdfplumber
python-dotenv
